Realign clone tag behavior and simplify all the clone tags.
