+++ src/cards/pathfinders/LobbyHalls.ts
@@ -38,8 +38,7 @@ export class LobbyHalls extends Card implements IProjectCard, ICloneTagCard {
   }
 
   public play(player: Player) {
-    player.game.defer(new DeclareCloneTag(player, this, 'Select the clone tag\'s new tag',
-      (tag) => PathfindersExpansion.raiseTrack(tag, player, 1)));
+    player.game.defer(new DeclareCloneTag(player, this, (tag) => PathfindersExpansion.raiseTrack(tag, player, 1)));
     player.adjustProduction(this.productionBox);
     const turmoil = Turmoil.getTurmoil(player.game);
     if (turmoil.getAvailableDelegateCount(player.id, 'reserve') > 0) {
