+++ tests/pathfinders/DeclareCloneTag.spec.ts
@@ -6,20 +6,25 @@ import {OrOptions} from '../../src//inputs/OrOptions';
 import {SelectOption} from '../../src//inputs/SelectOption';
 import {TestPlayer} from '../../tests/TestPlayer';
 import {getTestPlayer, newTestGame} from '../../tests/TestGame';
+import {TestingUtils} from '../TestingUtils';
+import {Game} from '../../src/Game';
+import {CrewTraining} from '../../src/cards/pathfinders/CrewTraining';
+import {MartianZoo} from '../../src/cards/colonies/MartianZoo';
 
 describe('DeclareCloneTag', function() {
   let player: TestPlayer;
+  let game: Game;
   let card: LobbyHalls;
   let tag: Tags;
 
   beforeEach(function() {
-    const game = newTestGame(1);
+    game = newTestGame(1);
     card = new LobbyHalls();
     player = getTestPlayer(game, 0);
   });
 
   it('sanity', function() {
-    const action = new DeclareCloneTag(player, card, '', (t) => tag = t);
+    const action = new DeclareCloneTag(player, card, (t) => tag = t);
 
     const options = action.execute();
     expect(options).instanceOf(OrOptions);
@@ -45,7 +50,7 @@ describe('DeclareCloneTag', function() {
     const game = newTestGame(1, {venusNextExtension: true, moonExpansion: true});
     player = getTestPlayer(game, 0);
 
-    const action = new DeclareCloneTag(player, card, '', (t) => tag = t);
+    const action = new DeclareCloneTag(player, card, (t) => tag = t);
 
     const options = action.execute();
     const orOptions = options.options as Array<SelectOption>;
@@ -59,5 +64,30 @@ describe('DeclareCloneTag', function() {
     orOptions[4].cb();
     expect(card.cloneTag).eq(Tags.MOON);
     expect(tag).eq(Tags.MOON);
+  });
+
+
+  it('When playing a card with a clone tag, onCardPlayed uses resulting tag', () => {
+    const crewTraining = new CrewTraining(); // Has two clone tags
+    const martianZoo = new MartianZoo(); // When you play an Earth tag, place an animal here
+    player.cardsInHand = [crewTraining];
+    player.playedCards = [martianZoo];
+
+    player.playCard(crewTraining);
+
+    expect(martianZoo.resourceCount).eq(0);
+
+
+    const action = TestingUtils.cast(DeclareCloneTag, game.deferredActions.pop());
+    const options = TestingUtils.cast(OrOptions, action!.execute());
+
+    expect(options.options[0].title).to.match(/earth/);
+
+    options.options[0].cb();
+
+    TestingUtils.runAllActions(game);
+
+    expect(crewTraining.tags).deep.eq([Tags.EARTH, Tags.EARTH]);
+    expect(martianZoo.resourceCount).eq(2);
   });
 });
